Fixed everything on this side - casted some map editor doubles to Integers because there was no type checking on parsing into strings
+Fixed map editor with Int casting
+ FInsihed up the DAO pattern stuff probably

+++ src/main/java/edu/wpi/cs3733/c22/teamC/PathFinding/PathFinder.java
@@ -8,7 +8,6 @@ import edu.wpi.cs3733.c22.teamC.Databases.requests.MaintenanceRequest;
 import edu.wpi.cs3733.c22.teamC.PathFinding.AStarImpl.AStar;
 import edu.wpi.cs3733.c22.teamC.PathFinding.AStarImpl.Node;
 import edu.wpi.cs3733.c22.teamC.SQLMethods.LocationQuery;
-import edu.wpi.cs3733.c22.teamC.SQLMethods.requests.MaintenanceRequestQuery;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.*;
@@ -43,8 +42,8 @@ public class PathFinder {
     for (Location loc : locationList) {
       Node node =
           new Node(
-              Integer.parseInt(loc.get_xcoord()),
-              Integer.parseInt(loc.get_ycoord()),
+              (int) Double.parseDouble(loc.get_xcoord()),
+              (int) Double.parseDouble(loc.get_ycoord()),
               loc.get_zcoord(),
               loc.get_nodeID());
       nodeMap.put(loc.get_nodeID(), node);
